app/requirements.txt
Django==3.2.6
celery==5.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
redis==3.5.3

six>=1.16.0,<1.17.0
multiqc>=1.10.1,<1.11.0
pymdown-extensions>=8.2,<8.3
pysam>=0.18.0,<0.19.0
cutadapt>=3.4,<3.5
biopython>=1.79,<2.0
pathlib>=1.0.1,<1.1.0
pandas>=1.3.5,<1.4.0
ete3>=3.1.2,<3.2.0

gtfparse>=1.2.1,<1.3.0
